getSetting() returns null for a missing key; say so and cast at the users
The docblock promised `string|array`, but both return paths hand back null
when the row is absent -- the array form does it deliberately, with a
comment explaining that null rather than '' keeps the single-key and
multi-key shapes agreeing.

That mattered here. PHPStan read the annotation and called
UserSessions::compatWrites()'s `null === $set` branch dead, and the
obvious way to clear the error would have been to delete the check. The
check is load-bearing: without it an unset FOG_AGENT_USERSESSION_COMPAT
returns (bool)null, so the setting the docblock calls "Defaults to ON"
would have defaulted to off.

Correcting the annotation surfaced nine daemons assigning the value
straight into an `@var int` property. Cast at those sites, which is what
each already meant -- every one of them immediately compares `< 1`, and
null and '0' and 0 all took the disabled branch before, so no behavior
moves.

Also drop \LDAP\Connection from FOGLdap::$_ld. ldap_connect() returns
that object from PHP 8.1 but a resource on 7.4, and phpstan.neon pins
phpVersion.min at 70400, where the class does not exist.

+++ packages/web/src/Service/FOGItemScanner.php
@@ -210,7 +210,7 @@ abstract class FOGItemScanner extends FOGService
         try {
             // Re-read every pass: a daemon must notice the setting being
             // turned off without needing a restart.
-            self::$_scanOn = self::getSetting(
+            self::$_scanOn = (int) self::getSetting(
                 $this->d('prefix') . 'GLOBALENABLED'
             );
             if (self::$_scanOn < 1) {
